+++ recipes/etcd_cluster.rb
@@ -114,6 +114,25 @@ if etcd_servers.find { |server_etcd| server_etcd['fqdn'] == node['fqdn'] }
 
   package 'etcd'
 
+  if node['cookbook-openshift3']['deploy_containerized']
+    execute 'Pull ETCD docker image' do
+      command "docker pull #{node['cookbook-openshift3']['openshift_docker_etcd_image']}"
+      not_if "docker images  | grep #{node['cookbook-openshift3']['openshift_docker_etcd_image']}"
+    end
+
+    template "/etc/systemd/system/#{node['cookbook-openshift3']['etcd_service_name']}.service" do
+      source 'service_etcd-containerized.service.erb'
+      notifies :run, 'execute[daemon-reload]', :immediately
+    end
+
+    ruby_block 'Mask ETCD service' do
+      block do
+        Mixlib::ShellOut.new('systemctl mask etcd').run_command
+      end
+      action :nothing
+    end
+  end
+
   remote_file "Retrieve certificate from ETCD Master[#{etcd_servers.first['fqdn']}]" do
     path "#{node['cookbook-openshift3']['etcd_conf_dir']}/etcd-#{node['fqdn']}.tgz"
     source "http://#{etcd_servers.first['ipaddress']}:#{node['cookbook-openshift3']['httpd_xfer_port']}/etcd/generated_certs/etcd-#{node['fqdn']}.tgz"
@@ -156,7 +175,7 @@ if etcd_servers.find { |server_etcd| server_etcd['fqdn'] == node['fqdn'] }
 
   template "#{node['cookbook-openshift3']['etcd_conf_dir']}/etcd.conf" do
     source 'etcd.conf.erb'
-    notifies :restart, 'service[etcd]', :immediately
+    notifies :restart, 'service[etcd-service]', :immediately
     variables lazy {
       {
         etcd_servers: etcd_servers,
@@ -165,7 +184,12 @@ if etcd_servers.find { |server_etcd| server_etcd['fqdn'] == node['fqdn'] }
     }
   end
 
-  service 'etcd' do
+  service 'etcd-service' do
     action [:start, :enable]
+  end
+
+  cookbook_file '/etc/profile.d/etcdctl.sh' do
+    source 'etcdctl.sh'
+    mode '0755'
   end
 end