+++ recipes/master.rb
@@ -9,6 +9,43 @@ master_servers = node['cookbook-openshift3']['master_servers']
 include_recipe 'cookbook-openshift3::etcd_cluster'
 
 if master_servers.find { |server_master| server_master['fqdn'] == node['fqdn'] }
+  if node['cookbook-openshift3']['deploy_containerized']
+    execute 'Pull CLI docker image' do
+      command "docker pull #{node['cookbook-openshift3']['openshift_docker_cli_image']}:#{node['cookbook-openshift3']['openshift_docker_image_version']}"
+      not_if "docker images  | grep #{node['cookbook-openshift3']['openshift_docker_cli_image']}.*#{node['cookbook-openshift3']['openshift_docker_image_version']}"
+    end
+
+    bash 'Add CLI to master(s)' do
+      code <<-EOH
+        docker create --name temp-cli ${DOCKER_IMAGE}:${DOCKER_TAG}
+        docker cp temp-cli:/usr/bin/openshift /usr/local/bin/openshift
+        docker rm temp-cli
+        EOH
+      environment(
+        'DOCKER_IMAGE' => node['cookbook-openshift3']['openshift_docker_master_image'],
+        'DOCKER_TAG' => node['cookbook-openshift3']['openshift_docker_image_version']
+      )
+      not_if { ::File.exist?('/usr/local/bin/openshift') }
+    end
+
+    %w(oadm oc kubectl).each do |client_symlink|
+      link "/usr/local/bin/#{client_symlink}" do
+        to '/usr/local/bin/openshift'
+        link_type :hard
+      end
+    end
+
+    execute 'Add bash completion for oc' do
+      command '/usr/local/bin/oc completion bash > /etc/bash_completion.d/oc'
+      not_if { ::File.exist?('/etc/bash_completion.d/oc') }
+    end
+
+    execute 'Pull MASTER docker image' do
+      command "docker pull #{node['cookbook-openshift3']['openshift_docker_master_image']}:#{node['cookbook-openshift3']['openshift_docker_image_version']}"
+      not_if "docker images  | grep #{node['cookbook-openshift3']['openshift_docker_master_image']}.*#{node['cookbook-openshift3']['openshift_docker_image_version']}"
+    end
+  end
+
   package node['cookbook-openshift3']['openshift_service_type'] do
     version node['cookbook-openshift3'] ['ose_version'] unless node['cookbook-openshift3']['ose_version'].nil?
     not_if { node['cookbook-openshift3']['deploy_containerized'] }